feat: add toggle option "Fast sync" to assumeutreexo

+++ app/src/main/java/com/github/jvsena42/mandacaru/domain/model/florestaRPC/response/GetBlockchainInfoResponse.kt
@@ -47,7 +47,7 @@ data class Result(
     @SerializedName("latest_work")
     val latestWork: String,
     @SerializedName("leaf_count")
-    val leafCount: Int,
+    val leafCount: Long,
     @SerializedName("progress")
     val progress: Float,
     @SerializedName("root_count")
